Renamed translate package to openioc2stix.
