+++ openioc2stix/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (c) 2015, The MITRE Corporation. All rights reserved.
+# See LICENSE.txt for complete terms.
