+++ openioc2stix/openioc.py
