+++ openioc2stix/translate.py
