+++ openioc2stix/utils.py
