+++ openioc2stix/xml.py
